+++ poc/poc_chat/screens/quit_screen.py
@@ -1,0 +1,26 @@
+# Copyright @ https://textual.textualize.io/guide/screens/#__tabbed_7_1
+
+from textual.app import ComposeResult
+from textual.containers import Grid
+from textual.screen import ModalScreen
+from textual.widgets import Label, Button
+
+
+class QuitScreen(ModalScreen[bool]):
+    """Screen with a dialog to quit."""
+
+    CSS_PATH = "quit_screen.tcss"
+
+    def compose(self) -> ComposeResult:
+        yield Grid(
+            Label("Are you sure you want to quit?", id="question"),
+            Button("Quit", variant="error", id="quit"),
+            Button("Cancel", variant="primary", id="cancel"),
+            id="dialog",
+        )
+
+    def on_button_pressed(self, event: Button.Pressed) -> None:
+        if event.button.id == "quit":
+            self.dismiss(True)
+        else:
+            self.dismiss(False)
