+++ poc/poc_chat/widgets/chats_list.py
@@ -1,0 +1,13 @@
+#  Copyright (c) 2024. Sergii Nechuiviter
+
+from textual.widgets import Placeholder
+
+
+class ChatsList(Placeholder):
+    DEFAULT_CSS = """
+    ChatsList {
+        width: 25vw;
+        dock: left;
+        border: solid white;
+    }
+    """
